# openapi.yaml
openapi: 3.0.0
info:
  title: 'LookDesk API'
  version: 1.0.0
servers:
  -
    url: 'https://api.lookdesk.ai'
    description: 'Production API'
paths:
...
  /api/enterprise-requests:
    post:
      tags:
        - Enterprise
      summary: 'Send enterprise request'
      operationId: fcda5b140756c4b0fb1bd3f12cc7af45
      requestBody:
        required: true
        content:
          application/json:
            schema: {  }
            example:
              name: 'John Doe'
              email: john@company.com
              employees_range: 101-200
              requirements: 'Need SSO and custom integrations'
      responses:
        '200':
          description: 'Request sent'
        '422':
          description: 'Validation error'
  /api/admin/enterprise-requests:
    get:
      tags:
        - 'Admin Enterprise'
      summary: 'Get enterprise requests list'
      operationId: 84eeedfa931911117aa816cadca7ce3e
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
        -
          name: per_page
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 'Paginated list'
        '403':
          description: Forbidden
      security:
        -
          sanctum: []
...
  /api/roles:
    get:
      tags:
        - Roles
      summary: 'Get all roles'
      operationId: ca86f0a206144dfbd695b749b4b28f9d
      responses:
        '200':
          description: 'Roles list'
      security:
        -
          sanctum: []
    post:
      tags:
        - Roles
      summary: 'Create new role'
      operationId: 341b8b162b61267c7bd187500aea5bc4
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              properties:
                name:
                  type: string
              type: object
      responses:
        '201':
          description: 'Role created'
      security:
        -
          sanctum: []
  '/api/roles/{id}':
    get:
      tags:
        - Roles
      summary: 'Get role by ID'
      operationId: 15c404caf238f1861c81f68205d328e5
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'Role details'
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    put:
      tags:
        - Roles
      summary: 'Update role'
      operationId: c73ab816ae8fb7c34a995d37ba577082
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
              type: object
      responses:
        '200':
          description: Updated
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Roles
      summary: 'Delete role'
      operationId: ffa6ef115d86b00cc6c54a6210019f9c
      responses:
        '200':
          description: Deleted
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
...
components:
...
  securitySchemes:
    sanctum:
      type: http
      bearerFormat: JWT
      scheme: bearer
...